backend/requirements.txt
pymongo
bcrypt
"fastapi[standard]"
python-socketio
fastapi